+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/ViewStyleCustomization.kt
@@ -168,6 +168,16 @@ class ViewStyleCustomization {
      */
     var locationPuck: LocationPuck? = null
 
+    /**
+     * Map scalebar params.
+     * Use [defaultMapScalebarParams] to reset to default.
+     * NOTE: When `enabled`, the `scalebar` will always be added to the top start corner
+     * of the screen. Position of the `scalebar` using `NavigationView` cannot be changed at any
+     * given time. However, if you change the position using `MapView`, the behavior is undefined
+     * and you will be responsible to ensure the correct positioning based on other view overlays.
+     */
+    var mapScalebarParams: MapboxMapScalebarParams? = null
+
     companion object {
         /**
          * Default info panel peek height in pixels.
@@ -374,6 +384,12 @@ class ViewStyleCustomization {
                 R.drawable.mapbox_navigation_puck_icon,
             )
         )
+
+        /**
+         * Default map scalebar parameters.
+         */
+        fun defaultMapScalebarParams(context: Context): MapboxMapScalebarParams =
+            MapboxMapScalebarParams.Builder(context).build()
 
         private fun Context.defaultSpacing() =
             resources.getDimensionPixelSize(R.dimen.mapbox_actionList_spacing)