+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/coordinator/ScalebarPlaceholderBinder.kt
@@ -1,0 +1,34 @@
+package com.mapbox.navigation.dropin.coordinator
+
+import android.view.View
+import android.view.ViewGroup
+import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
+import com.mapbox.navigation.core.lifecycle.MapboxNavigationObserver
+import com.mapbox.navigation.dropin.NavigationViewContext
+import com.mapbox.navigation.dropin.R
+import com.mapbox.navigation.dropin.component.map.ScalebarPlaceholderComponent
+import com.mapbox.navigation.dropin.databinding.MapboxScalebarPlaceholderLayoutBinding
+import com.mapbox.navigation.ui.app.internal.extension.actionsFlowable
+import com.mapbox.navigation.ui.base.lifecycle.UIBinder
+import kotlinx.coroutines.flow.map
+
+@OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
+internal class ScalebarPlaceholderBinder(
+    private val context: NavigationViewContext,
+) : UIBinder {
+
+    override fun bind(viewGroup: ViewGroup): MapboxNavigationObserver {
+        val binding = inflateLayout(viewGroup)
+        return ScalebarPlaceholderComponent(
+            binding.scalebarPlaceholder,
+            context.styles.mapScalebarParams,
+            context.store.state
+        )
+    }
+
+    private fun inflateLayout(viewGroup: ViewGroup): MapboxScalebarPlaceholderLayoutBinding {
+        viewGroup.removeAllViews()
+        View.inflate(viewGroup.context, R.layout.mapbox_scalebar_placeholder_layout, viewGroup)
+        return MapboxScalebarPlaceholderLayoutBinding.bind(viewGroup)
+    }
+}